test: split oversized agent status unit tests for max-lines
